ch19 optimizations

+++ src/main/java/com/quaxt/mcc/asm/Memory.java
@@ -1,7 +1,7 @@
 package com.quaxt.mcc.asm;
 
-public record Memory(Reg reg, int offset) implements Operand {
-    public Operand plus(int offset) {
+public record Memory(Reg reg, long offset) implements Operand {
+    public Operand plus(long offset) {
         return new Memory(reg, this.offset + offset);
     }
 }
